Create SimonSays
This is the code I used to create the simon says game on App Lab. The coding language is Java Script so most of it is usable. The only thing that you really need to edit for yourself would be all the functions and programming that control the User Interface such as start screens, play sounds etc. Essentially you can turn it into a text game if you like, or using HTML you can convert that code. For example: instead of "blue button" property, which is a square icon from App Lab, you can create a html block and call it blue button, then you can change and edit properties and etc. You can also still keep the change color functions by using RGB property of that block. It's a good foundation, you just gotta put in the work of understanding and being able to convert some stuff based on what software you would like to use to make this game. I'm sure you can do it, good luck!